ice: move gathering to application
ref: https://github.com/creytiv/re/wiki/ICE-Stacks-merge
