+++ src/ice/cand.c
@@ -258,6 +258,18 @@ struct ice_cand *icem_lcand_find_checklist(const struct icem *icem,
 }
 
 
+struct ice_cand *icem_lcand_base(struct ice_cand *lcand)
+{
+	return lcand ? lcand->base : NULL;
+}
+
+
+const struct sa *icem_lcand_addr(const struct ice_cand *cand)
+{
+	return cand ? &cand->addr : NULL;
+}
+
+
 int icem_cands_debug(struct re_printf *pf, const struct list *lst)
 {
 	struct le *le;
